Rename PlaylistManager to Playlist.
