+++ lib/itunes-playlist-to-usb/playlist.rb
